Add minion manager skeleton.
