+++ setup.cfg
@@ -30,6 +30,7 @@ console_scripts =
     coriolis-worker = coriolis.cmd.worker:main
     coriolis-replica-cron = coriolis.cmd.replica_cron:main
     coriolis-scheduler= coriolis.cmd.scheduler:main
+    coriolis-minion-manager= coriolis.cmd.minion_manager:main
     coriolis-dbsync = coriolis.cmd.db_sync:main
 
 [wheel]
